clean up add 33 mode
