--- FridaLoader/app/build.gradle
@@ -1,48 +1,0 @@
-apply plugin: 'com.android.application'
-
-android {
-    compileSdkVersion 29
-    defaultConfig {
-        applicationId "com.dns.fridaloader"
-        minSdkVersion 16
-        targetSdkVersion 29
-        versionCode 1
-        versionName "1.0"
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-    }
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
-        }
-    }
-    compileOptions {
-        sourceCompatibility = '1.8'
-        targetCompatibility = '1.8'
-    }
-}
-repositories {
-    maven { url 'https://jitpack.io' }
-}
-
-dependencies {
-    implementation 'com.github.sabujak-sabujak:RoundedButton:v0.1.1'
-}
-dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.mindorks.android:prdownloader:+'
-    implementation 'com.muddzdev:styleabletoast:+'
-
-
-
-
-    implementation 'com.crowdfire.cfalertdialog:cfalertdialog:1.1.0'
-
-    implementation group: 'org.tukaani', name: 'xz', version: '1.8'
-   // implementation 'com.android.support:appcompat-v7:+'
-    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
-    testImplementation 'junit:junit:4.12'
-  //  androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.0-alpha4'
-}
-
