+++ FridaLoader/src/main/java/com/dns/fridaloader/MainApp.java
@@ -6,14 +6,10 @@ import com.downloader.PRDownloader;
 import com.downloader.PRDownloaderConfig;
 
 public class MainApp extends Application {
-
     @Override
     public void onCreate() {
         super.onCreate();
-        PRDownloaderConfig config = PRDownloaderConfig.newBuilder()
-                .setDatabaseEnabled(true)
-                .build();
+        PRDownloaderConfig config = PRDownloaderConfig.newBuilder().setDatabaseEnabled(true).build();
         PRDownloader.initialize(this, config);
     }
-
 }
