+++ FridaLoader/src/main/java/com/dns/fridaloader/Util.java
@@ -1,23 +1,20 @@
 package com.dns.fridaloader;
 
-
 import android.content.Context;
 import android.os.Environment;
+
 import androidx.core.content.ContextCompat;
 
 import java.io.File;
 import java.util.Locale;
 
 public final class Util {
-
     private Util() {
-
     }
 
     public static String getRootDirPath(Context context) {
         if (Environment.MEDIA_MOUNTED.equals(Environment.getExternalStorageState())) {
-            File file = ContextCompat.getExternalFilesDirs(context.getApplicationContext(),
-                    null)[0];
+            File file = ContextCompat.getExternalFilesDirs(context.getApplicationContext(),null)[0];
             return file.getAbsolutePath();
         } else {
             return context.getApplicationContext().getFilesDir().getAbsolutePath();
@@ -28,8 +25,7 @@ public final class Util {
         return getBytesToMBString(currentBytes) + "/" + getBytesToMBString(totalBytes);
     }
 
-    private static String getBytesToMBString(long bytes){
+    private static String getBytesToMBString(long bytes) {
         return String.format(Locale.ENGLISH, "%.2fMb", bytes / (1024.00 * 1024.00));
     }
-
 }
