+++ FridaLoader/src/main/res/mipmap-xxxhdpi/ic_launcher.png
