+++ FridaLoader/src/main/res/values/colors.xml
