+++ FridaLoader/src/main/res/values/strings.xml
