+++ FridaLoader/src/main/res/values/styles.xml
