+++ local.properties
@@ -1,0 +1,9 @@
+## This file must *NOT* be checked into Version Control Systems,
+# as it contains information specific to your local configuration.
+#
+# Location of the SDK. This is only used by Gradle.
+# For customization when using a Version Control System, please read the
+# header note.
+#Thu Jan 19 17:25:40 CET 2023
+sdk.dir=C\:\\Users\\msntc\\AppData\\Local\\Android\\Sdk
+ndk.dir=C\:\\Users\\msntc\\AppData\\Local\\Android\\Sdk\\ndk\\21.4.7075529
